init: Reinitialize b68-app to Typescript 📲
