+++ packages/app/assets/adaptive-icon.png
