+++ packages/app/assets/favicon.png
