+++ packages/app/babel.config.js
@@ -1,4 +1,5 @@
-export default function (api) {
+// eslint-disable-next-line no-undef
+module.exports = function (api) {
     api.cache(true)
     return {
         presets: ['babel-preset-expo'],
