+++ packages/app/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "b68-app",
+    "name": "app",
     "version": "1.0.0",
     "main": "node_modules/expo/AppEntry.js",
     "scripts": {
@@ -15,7 +15,9 @@
         "react-native": "0.71.8"
     },
     "devDependencies": {
-        "@babel/core": "^7.20.0"
+        "@babel/core": "^7.20.0",
+        "@types/react": "~18.0.14",
+        "typescript": "^4.9.4"
     },
     "private": true
 }
