+++ packages/app/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "expo/tsconfig.base",
+    "compilerOptions": {
+        "strict": true
+    }
+}
