--- packages/b68-app/App.js
@@ -1,20 +1,0 @@
-import { StatusBar } from 'expo-status-bar'
-import { StyleSheet, Text, View } from 'react-native'
-
-export default function App() {
-    return (
-        <View style={styles.container}>
-            <Text>B68 App</Text>
-            <StatusBar style="auto" />
-        </View>
-    )
-}
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#fff',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-})
